requirements.txt
asn1crypto==0.23.0
attrs==17.3.0
Automat==0.6.0
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
constantly==15.1.0
cryptography==2.1.3
cssselect==1.0.1
fake-useragent==0.1.8
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
lxml==4.1.1
mysqlclient==1.3.12
olefile==0.44
parsel==1.2.0
Pillow==4.3.0
pyasn1==0.3.7
pyasn1-modules==0.1.5
pycparser==2.18
PyDispatcher==2.0.5
pymongo==3.5.1
pyOpenSSL==17.3.0
queuelib==1.4.2
redis==2.10.6
requests==2.18.4
Scrapy==1.4.0
scrapy-splash==0.7.2
selenium==3.8.0
service-identity==17.0.0
six==1.11.0
SQLAlchemy==1.1.15
Twisted==17.9.0
urllib3==1.22
w3lib==1.18.0
zope.interface==4.4.3
